Each region (header, main, left, right, footer) was sharing the same collection for Modules.
When a module was added through the command view, it would render 5 times, once in each region (each region was binded to the "add" event of the collection).

From Backbone.js docs:
"Remember that in JavaScript, objects are passed by reference, so if you include an object as a default value, it will be shared among all instances. Instead, define defaults as a function."

git-svn-id: https://aux1.edublogs.org/repos/incsub/trunk/themes/upfront@9040 f53fe46e-8c02-459b-9fdd-93e75771b278