Fix issues with global regions
Adding global regions when needed was kinda messy and would not yield
good results because even when new layout was created it was done by
trying to load it and if loading failed than it was created. This led to
some logic fails.
Also layout loading itself was problematic since it was not loading
regions from theme files correctly and than it triggered creation of
layouts even if they existed in theme and since creation was done by
loading... that was kinda wreck.
So, this fix is not just about global regions but generaly about
loading/creation of layouts.