Refactor grid CSS, fix region z-index
Note: Now the grid CSS is the same for both front-end and editor, which means I have deprecated the "light" grid CSS. This is to allow a more consistent result with front-end and editor. The size is a concern at first, with front-end one load at around 89kB, but when gzipped, this reduced to just 16kB.  This grid CSS is more robust with correct overriding, which the lighter CSS lack of (and yes, the grid CSS is already optimized as it is, can't see a way to minimalize it more at the moment). So it's a good trade off personally and with gzipped size, I'm happy. :)