Fix bug where layout was alternating between _dev and normal version
Issue:
When calling the function Upfront_Layout::from_entity_ids() multiple times, then
the first result would be the _dev version, second result the non-dev version,
then again _dev version, etc.

Caused by condition in line 356, that would only change the storage key to _dev
when it was NOT _dev already - so the second call would not go into that
condition and as a result the storage key would be set to empty value "", which
resets the key to default non-dev key.