Add content refiltering for MP compat layer
This should fix conflict that can happen between
MarketPress, our MP compat layer and a third plugin
when it's introduced in the mix.

The issue manifests when the third plugin triggers
content filtering ahead of time, before we're doing our
content filtering in post data element. When these
conditions are met, MP removes the filter way too soon
(i.e. before we show anything) and we end up with a bare
post-like content with no additional processing.

The fix attempts to detect this condition, and to re-bind
the MP content filtering just in time for it to kick in
when the post data element content is being rendered.